Transform Main.extractParent into implicit class (#14)
